added beta 1d generator
